requirements.txt
boto3~=1.20.0
chaostoolkit-aws>=0.13.0,<1.0.0
chaostoolkit-cloud-foundry>=0.7.1,<1.0.0
chaostoolkit-istio>=0.1.4,<1.0.0
chaostoolkit-kubernetes>=0.24.0,<1.0.0
chaostoolkit-lib>=0.22.2,<1.0.0
chaostoolkit-prometheus>=0.3.0,<1.0.0
chaostoolkit-spring<=0.2.0,<1.0.0
croniter==0.3.29
django~=2.2.0
django-cors-headers==3.1.1
djangorestframework~=3.11.2
drf-yasg~=1.20.0
gunicorn==20.0.4
huey~=2.3.0
jinja2==2.11.3
jsonpath-ng==1.4.3
MarkupSafe==2.0.1  # 2.1.0 is incompatible with jinja2 2.11.3
mysql-connector-python~=8.0.0
peewee==3.9.3
pycryptodome~=3.10.0
python-daemon==2.1.2
python-jose[pycryptodome]~=3.3.0
python-logstash==0.4.6
requests~=2.26.0
snaptime==0.2.4
urllib3==1.26.5
